refactor: move unordered-list to ui
